Refactor: 로그인 네비게이션
