+++ feature/signin/src/main/java/com/acon/feature/signin/screen/SignInRoute.kt
@@ -1,4 +1,4 @@
-package com.acon.acon.navigation.route
+package com.acon.feature.signin.screen
 
 import kotlinx.serialization.Serializable
